+++ govtool/frontend/src/consts/governanceActionFields.ts
@@ -7,6 +7,11 @@ import {
 } from "@/types/governanceAction";
 import { bech32 } from "bech32";
 
+export const CIP_100 =
+  "https://github.com/cardano-foundation/CIPs/blob/master/CIP-0100/README.md#";
+export const CIP_108 =
+  "https://github.com/cardano-foundation/CIPs/blob/master/CIP-0108/README.md#";
+
 const sharedGovernanceActionFields: SharedGovernanceActionFieldSchema = {
   title: {
     component: GovernanceActionField.Input,
@@ -120,3 +125,68 @@ export const GOVERNANCE_ACTION_FIELDS: GovernanceActionFields = {
     },
   },
 } as const;
+
+const commonContext = {
+  "@language": "en-us",
+  CIP100:
+    "https://github.com/cardano-foundation/CIPs/blob/master/CIP-0100/README.md#",
+  CIP108:
+    "https://github.com/cardano-foundation/CIPs/blob/master/CIP-0108/README.md#",
+  hashAlgorithm: "CIP100:hashAlgorithm",
+  body: {
+    "@id": "CIP108:body",
+    "@context": {
+      references: {
+        "@id": "CIP108:references",
+        "@container": "@set" as "@set",
+        "@context": {
+          GovernanceMetadata: "CIP100:GovernanceMetadataReference",
+          Other: "CIP100:OtherReference",
+          label: "CIP100:reference-label",
+          uri: "CIP100:reference-uri",
+          referenceHash: {
+            "@id": "CIP108:referenceHash",
+            "@context": {
+              hashDigest: "CIP108:hashDigest",
+              hashAlgorithm: "CIP100:hashAlgorithm",
+            },
+          },
+        },
+      },
+      title: "CIP108:title",
+      abstract: "CIP108:abstract",
+      motivation: "CIP108:motivation",
+      rationale: "CIP108:rationale",
+    },
+  },
+  authors: {
+    "@id": "CIP100:authors",
+    "@container": "@set" as "@set",
+    "@context": {
+      name: "http://xmlns.com/foaf/0.1/name",
+      witness: {
+        "@id": "CIP100:witness",
+        "@context": {
+          witnessAlgorithm: "CIP100:witnessAlgorithm",
+          publicKey: "CIP100:publicKey",
+          signature: "CIP100:signature",
+        },
+      },
+    },
+  },
+};
+
+export const GOVERNANCE_ACTION_CONTEXTS = {
+  [GovernanceActionType.Info]: commonContext,
+  [GovernanceActionType.Treasury]: {
+    ...commonContext,
+    body: {
+      ...commonContext.body,
+      "@context": {
+        ...commonContext.body["@context"],
+        amount: "CIP108:amount",
+        receivingAddress: "CIP108:receivingAddress",
+      },
+    },
+  },
+};