+++ govtool/frontend/src/hooks/forms/useCreateGovernanceActionForm.ts
@@ -5,6 +5,10 @@ import {
 import { useCallback, useState } from "react";
 import { useFormContext } from "react-hook-form";
 
+import { CIP_100, CIP_108, GOVERNANCE_ACTION_CONTEXTS } from "@consts";
+
+import * as jsonld from "jsonld";
+
 export type CreateGovernanceActionValues = {
   links?: { link: string }[];
   storeData?: boolean;
@@ -33,21 +37,47 @@ export const useCreateGovernanceActionForm = () => {
     reset,
   } = useFormContext<CreateGovernanceActionValues>();
 
+  const govActionType = watch("governance_action_type");
+
   // TODO: To be moved to utils
-  const generateJsonBody = (data: CreateGovernanceActionValues) => {
-    const filteredData = Object.entries(data).filter(
-      ([key]) => !Object.keys(defaulCreateGovernanceActionValues).includes(key)
-    );
+  const generateJsonBody = async (data: CreateGovernanceActionValues) => {
+    const filteredData = Object.entries(data)
+      .filter(
+        ([key]) =>
+          !Object.keys(defaulCreateGovernanceActionValues).includes(key) &&
+          key !== "governance_action_type"
+      )
+      .map(([key, value]) => {
+        return [CIP_108 + key, value];
+      });
+
     const references = (data as CreateGovernanceActionValues).links
       ?.filter((link) => link.link)
       .map((link) => {
-        // TODO: Label isnt available and hardcoded Other for type
-        return { "@type": "Other", label: "Testlabel", uri: link.link };
+        return {
+          [`@type`]: "Other",
+          [`${CIP_100}reference-label`]: "Label",
+          [`${CIP_100}reference-uri`]: link.link,
+        };
       });
-    const body = { ...Object.fromEntries(filteredData), references };
 
-    const jsonStr = JSON.stringify(body);
-    return jsonStr;
+    const body = {
+      ...Object.fromEntries(filteredData),
+      [`${CIP_108}references`]: references,
+    };
+
+    const doc = {
+      [`${CIP_108}body`]: body,
+      [`${CIP_100}hashAlgorithm`]: "blake2b-256",
+      [`${CIP_100}authors`]: [],
+    };
+
+    const json = await jsonld.compact(
+      doc,
+      GOVERNANCE_ACTION_CONTEXTS[govActionType as GovernanceActionType]
+    );
+
+    return json;
   };
 
   const onSubmit = useCallback(async (data: CreateGovernanceActionValues) => {